# contracts/openapi/meeting-service.openapi.yaml
openapi: 3.1.0
info:
  title: Meeting Service Internal API
  version: 1.0.0
  description: >-
    Contract called by the eCabinet BFF. Meeting Service owns runtime,
    transcripts, minutes, exports and realtime state; it never queries the
    eCabinet database. Meeting Service creates immutable evidence snapshots
    from its persisted final transcript, then calls the separate Meeting AI
    contract (`/internal/v1/sessions/{runtime_session_id}/analyze`).
servers:
  - url: http://meeting-service:8002
    description: Docker internal network
security:
  - ServiceKey: []
    ActorClaims: []
paths:
...
  /internal/v1/meetings/{meeting_id}/runtime:
    parameters:
      - $ref: "#/components/parameters/MeetingId"
    post:
      operationId: createMeetingRuntime
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../schemas/meeting-snapshot.schema.json"
      responses:
        "201":
          $ref: "#/components/responses/RuntimeResponse"
        "403":
          $ref: "#/components/responses/ProblemResponse"
        "409":
          $ref: "#/components/responses/ProblemResponse"
        "503":
          $ref: "#/components/responses/ProblemResponse"
  /internal/v1/runtimes/{runtime_session_id}/stop:
    parameters:
      - $ref: "#/components/parameters/RuntimeSessionId"
    post:
      operationId: stopMeetingRuntime
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          $ref: "#/components/responses/RuntimeResponse"
        "403":
          $ref: "#/components/responses/ProblemResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
        "409":
          $ref: "#/components/responses/ProblemResponse"
  /internal/v1/meetings/{meeting_id}/status:
    parameters:
      - $ref: "#/components/parameters/MeetingId"
    get:
      operationId: getMeetingRuntimeStatus
      responses:
        "200":
          $ref: "#/components/responses/RuntimeResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
  /internal/v1/meetings/{meeting_id}/snapshot:
    parameters:
      - $ref: "#/components/parameters/MeetingId"
    put:
      operationId: updateMeetingRuntimeSnapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "../schemas/meeting-snapshot.schema.json"
      responses:
        "200":
          $ref: "#/components/responses/SnapshotResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
        "409":
          $ref: "#/components/responses/ProblemResponse"
        "422":
          $ref: "#/components/responses/ProblemResponse"
  /internal/v1/meetings/{meeting_id}/tokens:
    parameters:
      - $ref: "#/components/parameters/MeetingId"
    post:
      operationId: issueMeetingLiveKitToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [runtime_session_id, user_id, device_id]
              properties:
                runtime_session_id:
                  type: string
                  format: uuid
                user_id:
                  type: string
                  format: uuid
                device_id:
                  type: string
                  minLength: 1
                  maxLength: 160
                permissions:
                  type: array
                  items: {type: string}
      responses:
        "200":
          description: LiveKit token. eCabinet signs the separate Socket.IO runtime token.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [runtime_session_id, livekit_url, room, identity, can_publish, token, expires_at]
                properties:
                  runtime_session_id: {type: string, format: uuid}
                  livekit_url: {type: string, format: uri}
                  room: {type: string}
                  identity: {type: string}
                  can_publish: {type: boolean}
                  token: {type: string}
                  expires_at: {type: string, format: date-time}
        "403":
          $ref: "#/components/responses/ProblemResponse"
        "409":
          $ref: "#/components/responses/ProblemResponse"
        "422":
          $ref: "#/components/responses/ProblemResponse"
        "503":
          $ref: "#/components/responses/ProblemResponse"
  /internal/v1/meetings/{meeting_id}/transcript:
    parameters:
      - $ref: "#/components/parameters/MeetingId"
    get:
      operationId: listMeetingTranscripts
      responses:
        "200":
          description: Persisted final transcript state.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [meeting_id, segments]
                properties:
                  meeting_id: {type: string, format: uuid}
                  segments:
                    type: array
                    items: {type: object}
  /internal/v1/meetings/{meeting_id}/minutes:
    parameters:
      - $ref: "#/components/parameters/MeetingId"
    get:
      operationId: getMeetingMinutes
      responses:
        "200":
          $ref: "#/components/responses/MinutesResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
    patch:
      operationId: editMeetingMinutes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [base_revision, document]
              properties:
                base_revision: {type: integer, minimum: 0}
                document:
                  $ref: "../schemas/minutes-document.schema.json"
      responses:
        "200":
          $ref: "#/components/responses/MinutesResponse"
        "409":
          $ref: "#/components/responses/ProblemResponse"
        "422":
          $ref: "#/components/responses/ProblemResponse"
  /internal/v1/meetings/{meeting_id}/minutes/analyze:
    parameters:
      - $ref: "#/components/parameters/MeetingId"
    get:
      operationId: getMeetingMinutesAnalysisStatus
      responses:
        "200":
          $ref: "#/components/responses/MinutesAnalysisResponse"
    post:
      operationId: requestMeetingMinutesAnalysis
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          $ref: "#/components/responses/MinutesAnalysisResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
        "409":
          $ref: "#/components/responses/ProblemResponse"
        "503":
          $ref: "#/components/responses/ProblemResponse"
  /internal/v1/meetings/{meeting_id}/minutes/auto-update/stop:
    parameters:
      - $ref: "#/components/parameters/MeetingId"
    post:
      operationId: stopMeetingMinutesAutoUpdate
      responses:
        "200":
          $ref: "#/components/responses/MinutesAnalysisResponse"
        "404":
          $ref: "#/components/responses/ProblemResponse"
  /internal/v1/meetings/{meeting_id}/minutes/review:
    parameters:
      - $ref: "#/components/parameters/MeetingId"
    post:
      operationId: submitMeetingMinutesForReview
      responses:
        "200":
          $ref: "#/components/responses/MinutesResponse"
        "409":
          $ref: "#/components/responses/ProblemResponse"
  /internal/v1/meetings/{meeting_id}/minutes/approve:
    parameters:
      - $ref: "#/components/parameters/MeetingId"
    post:
      operationId: approveMeetingMinutes
      responses:
        "200":
          $ref: "#/components/responses/MinutesResponse"
        "403":
          $ref: "#/components/responses/ProblemResponse"
        "409":
          $ref: "#/components/responses/ProblemResponse"
  /internal/v1/meetings/{meeting_id}/minutes/exports/docx:
    parameters:
      - $ref: "#/components/parameters/MeetingId"
    post:
      operationId: exportMeetingMinutesDocx
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                revision: {type: integer, minimum: 1}
                allow_draft: {type: boolean, default: false}
                created_by: {type: string, maxLength: 160}
      responses:
        "201":
          description: Export metadata created.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MinutesExport"}
        "403":
          $ref: "#/components/responses/ProblemResponse"
        "409":
          $ref: "#/components/responses/ProblemResponse"
  /internal/v1/meetings/{meeting_id}/minutes/exports/{export_id}:
    parameters:
      - $ref: "#/components/parameters/MeetingId"
      - name: export_id
        in: path
        required: true
        schema: {type: string, format: uuid}
    get:
      operationId: downloadMeetingMinutesExport
      responses:
        "200":
          description: DOCX export binary.
          content:
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema: {type: string, format: binary}
        "404": {$ref: "#/components/responses/ProblemResponse"}
  /internal/v1/meetings/{meeting_id}/minutes/exports/{export_id}/metadata:
    parameters:
      - $ref: "#/components/parameters/MeetingId"
      - name: export_id
        in: path
        required: true
        schema: {type: string, format: uuid}
    get:
      operationId: getMeetingMinutesExportMetadata
      responses:
        "200":
          description: DOCX export metadata.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MinutesExport"}
        "404": {$ref: "#/components/responses/ProblemResponse"}
  /internal/v1/meetings/{meeting_id}:
    parameters:
      - $ref: "#/components/parameters/MeetingId"
    delete:
      operationId: purgeMeetingRuntimeData
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Purge completed or already applied.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [meeting_id, status]
                properties:
                  meeting_id: {type: string, format: uuid}
                  status: {type: string, enum: [PURGED]}
                  runtime_rows_deleted: {type: integer, minimum: 0}
                  content_rows_deleted: {type: integer, minimum: 0}
                  ai_event_rows_deleted: {type: integer, minimum: 0}
  /internal/v1/enrollments/{user_id}:
    parameters:
      - $ref: "#/components/parameters/UserId"
    post:
      operationId: enrollMeetingSpeaker
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [display_name, audio]
              properties:
                display_name: {type: string, minLength: 1, maxLength: 160}
                audio: {type: string, format: binary}
      responses:
        "200": {description: Enrollment result.}
        "422": {$ref: "#/components/responses/ProblemResponse"}
    get:
      operationId: getMeetingSpeakerEnrollment
      responses:
        "200": {description: Enrollment profile status.}
        "404": {$ref: "#/components/responses/ProblemResponse"}
    delete:
      operationId: deleteMeetingSpeakerEnrollment
      responses:
        "204": {description: Profile removed or absent.}
components:
  securitySchemes:
    ServiceKey:
      type: apiKey
      in: header
      name: X-Service-Key
    ActorClaims:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    MeetingId:
      name: meeting_id
      in: path
      required: true
      schema: {type: string, format: uuid}
    RuntimeSessionId:
      name: runtime_session_id
      in: path
      required: true
      schema: {type: string, format: uuid}
    UserId:
      name: user_id
      in: path
      required: true
      schema: {type: string, format: uuid}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 8, maxLength: 160}
  schemas:
...
    MinutesExport:
      type: object
      additionalProperties: false
      required: [id, meeting_id, minutes_revision, minutes_status, format, storage_key, filename, content_type, size_bytes, checksum, created_at]
      properties:
        id: {type: string, format: uuid}
        meeting_id: {type: string, format: uuid}
        minutes_revision: {type: integer, minimum: 1}
        minutes_status: {type: string, enum: [DRAFT, REVIEWING, APPROVED, SUPERSEDED]}
        format: {type: string, enum: [docx]}
        storage_key: {type: string}
        filename: {type: string}
        content_type: {type: string}
        size_bytes: {type: integer, minimum: 0}
        checksum: {type: string}
        created_by: {type: [string, "null"]}
        created_at: {type: string, format: date-time}
...
  responses:
...
    RuntimeResponse:
      description: Runtime state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Runtime"
    SnapshotResponse:
      description: Updated participant and role snapshot.
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [meeting_id, runtime_session_id, snapshot_revision, status, snapshot]
            properties:
              meeting_id: {type: string, format: uuid}
              runtime_session_id: {type: string, format: uuid}
              snapshot_revision: {type: integer, minimum: 1}
              status:
                type: string
                enum: [STARTING, READY, RECORDING, STOPPING, COMPLETED, FAILED]
              snapshot:
                $ref: "../schemas/meeting-snapshot.schema.json"
    MinutesResponse:
      description: Minutes revision.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MinutesRevision"
    MinutesAnalysisResponse:
      description: Analysis request state. The evidence remains owned by Meeting Service.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MinutesAnalysis"
    ProblemResponse:
      description: Request failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"